+++ src/bin/TutorialD/Interpreter/Import/BasicExamples.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE CPP #-}
 --includes some hardcoded examples which can be imported even during safe evaluation (no file I/O)
 module TutorialD.Interpreter.Import.BasicExamples where
 import ProjectM36.DateExamples
@@ -6,10 +5,6 @@ import ProjectM36.Base
 import ProjectM36.Interpreter
 import ProjectM36.DatabaseContextExpr
 import TutorialD.Interpreter.Base
-
-#if !MIN_VERSION_megaparsec(6,0,0)
-import Text.Megaparsec.Text
-#endif
 
 data ImportBasicExampleOperator = ImportBasicDateExampleOperator
                                 deriving (Show)