+++ src/bin/TutorialD/Interpreter/RelationalExpr.hs
@@ -1,11 +1,6 @@
-{-# LANGUAGE CPP #-}
 module TutorialD.Interpreter.RelationalExpr where
 import Text.Megaparsec
-#if MIN_VERSION_megaparsec(7,0,0)
 import Control.Monad.Combinators.Expr
-#else
-import Text.Megaparsec.Expr
-#endif
 import ProjectM36.Base
 import ProjectM36.Interpreter
 import TutorialD.Interpreter.Base